Refactor: Remove unused NFCTag model and related code
