--- backend/ntags/models/__init__.py
@@ -1,6 +1,0 @@
-from .base import (
-    NFCTagManager, AbstractNFCTag, NFCTag,
-    NFCTagLinkManager, AbstractNFCTagLink, NFCTagLink
-)
-from .memory import NFCTagMemory
-from .scans import NFCTagScan
